code splitting and directory restructuring for CursorHandler plugin
