--- packages/shared/constants/helperCharacters.ts
@@ -1,4 +1,0 @@
-export const ZERO_WIDTH_SPACE = "\u200B";
-export const CARET = "\u2038";
-
-export const CURSOR_POSITION_HELPER_CHAR = CARET;
