+++ packages/shared/plugins/CursorHandler/core/utils/constants.ts
@@ -1,0 +1,14 @@
+export const ZERO_WIDTH_SPACE = "\u200B";
+export const CARET = "\u2038";
+
+export const CURSOR_PLACEHOLDER_CHAR = CARET;
+
+export enum CursorMovementDirection {
+  LEFT = "left",
+  RIGHT = "right",
+}
+
+export enum charSelectionOffset {
+  BEFORE = 0,
+  AFTER = 1,
+}
